27/08/2018 Por turnos:      La partida rápida tiene que matcher por las categorías seleccionadas.

+++ app/src/main/java/com/trivial/upv/android/fragment/PlayRealTimeFragment.java
@@ -96,7 +96,7 @@ public class PlayRealTimeFragment extends Fragment
             Game.listCategories.add(new String(category.getCategory()));
         }
 
-        Game.level = (long) (Math.pow(2, categoriesJSON.size()) - 1);
+        Game.level = (int) (Math.pow(2, categoriesJSON.size()) - 1);
     }
 
     @Override
